+++ tests/Extension/AssetRuntimeExtensionTest.php
@@ -14,6 +14,7 @@ namespace Berlioz\Package\Twig\Tests\Extension;
 
 use Berlioz\Core\Asset\Assets;
 use Berlioz\Package\Twig\Extension\AssetRuntimeExtension;
+use Berlioz\Router\Router;
 use PHPUnit\Framework\TestCase;
 use Twig\Error\RuntimeError;
 
@@ -23,6 +24,7 @@ class AssetRuntimeExtensionTest extends TestCase
 
     protected function setUp(): void
     {
+        unset($_SERVER['HTTP_X_FORWARDED_PREFIX']);
         $this->assets = new Assets(
             __DIR__ . '/data/manifest.json',
             __DIR__ . '/data/entrypoints.json',
@@ -31,7 +33,7 @@ class AssetRuntimeExtensionTest extends TestCase
 
     public function testAsset()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals('/assets/css/website.css', $extensionRuntime->asset('website.css'));
     }
@@ -41,13 +43,13 @@ class AssetRuntimeExtensionTest extends TestCase
         $this->expectException(RuntimeError::class);
         $this->expectExceptionMessage('Asset "fake.css" not found in manifest file');
 
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
         $extensionRuntime->asset('fake.css');
     }
 
     public function testEntryPoints()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals(
             '<link rel="stylesheet" href="/assets/css/website.css">' . PHP_EOL .
@@ -59,7 +61,7 @@ class AssetRuntimeExtensionTest extends TestCase
 
     public function testEntryPoints_withMultipleEntry()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals(
             '<link rel="stylesheet" href="/assets/css/website.css">' . PHP_EOL .
@@ -71,9 +73,24 @@ class AssetRuntimeExtensionTest extends TestCase
         );
     }
 
+    public function testEntryPoints_withMultipleEntryAndRouterOptions()
+    {
+        $_SERVER['HTTP_X_FORWARDED_PREFIX'] = '/prefix';
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router(['X-Forwarded-Prefix' => true]));
+
+        $this->assertEquals(
+            '<link rel="stylesheet" href="/prefix/assets/css/website.css">' . PHP_EOL .
+            '<link rel="stylesheet" href="/prefix/assets/css/admin.css">' . PHP_EOL .
+            '<script src="/prefix/assets/js/website.js"></script>' . PHP_EOL .
+            '<script src="/prefix/assets/js/vendor.js"></script>' . PHP_EOL .
+            '<script src="/prefix/assets/js/admin.js"></script>' . PHP_EOL,
+            $extensionRuntime->entryPoints(['website', 'admin'])
+        );
+    }
+
     public function testEntryPoints_withType()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals(
             '<script src="/assets/js/website.js"></script>' . PHP_EOL .
@@ -82,9 +99,21 @@ class AssetRuntimeExtensionTest extends TestCase
         );
     }
 
+    public function testEntryPoints_withTypeAndRouterOptions()
+    {
+        $_SERVER['HTTP_X_FORWARDED_PREFIX'] = '/prefix';
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router(['X-Forwarded-Prefix' => true]));
+
+        $this->assertEquals(
+            '<script src="/prefix/assets/js/website.js"></script>' . PHP_EOL .
+            '<script src="/prefix/assets/js/vendor.js"></script>' . PHP_EOL,
+            $extensionRuntime->entryPoints('website', 'js')
+        );
+    }
+
     public function testEntryPoints_withOptions()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals(
             '<link async defer attr="fake" data-first="value1" data-second="value2" data-third rel="stylesheet" href="/assets/css/website.css">' . PHP_EOL .
@@ -107,14 +136,14 @@ class AssetRuntimeExtensionTest extends TestCase
 
     public function testEntryPoints_notFound()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals('', $extensionRuntime->entryPoints('fake'));
     }
 
     public function testEntryPointsList()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals(
             ['css' => ['/assets/css/website.css'], 'js' => ['/assets/js/website.js', '/assets/js/vendor.js']],
@@ -124,7 +153,7 @@ class AssetRuntimeExtensionTest extends TestCase
 
     public function testEntryPointsList_withMultipleEntry()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals(
             ['/assets/js/website.js', '/assets/js/vendor.js', '/assets/js/admin.js'],
@@ -132,9 +161,20 @@ class AssetRuntimeExtensionTest extends TestCase
         );
     }
 
+    public function testEntryPointsList_withMultipleEntryAndRouterOptions()
+    {
+        $_SERVER['HTTP_X_FORWARDED_PREFIX'] = '/prefix';
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router(['X-Forwarded-Prefix' => true]));
+
+        $this->assertEquals(
+            ['/prefix/assets/js/website.js', '/prefix/assets/js/vendor.js', '/prefix/assets/js/admin.js'],
+            $extensionRuntime->entryPointsList(['website', 'admin'], 'js')
+        );
+    }
+
     public function testEntryPointsList_withType()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals(
             ['/assets/js/website.js', '/assets/js/vendor.js'],
@@ -142,9 +182,20 @@ class AssetRuntimeExtensionTest extends TestCase
         );
     }
 
+    public function testEntryPointsList_withTypeAndRouterOptions()
+    {
+        $_SERVER['HTTP_X_FORWARDED_PREFIX'] = '/prefix';
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router(['X-Forwarded-Prefix' => true]));
+
+        $this->assertEquals(
+            ['/prefix/assets/js/website.js', '/prefix/assets/js/vendor.js'],
+            $extensionRuntime->entryPointsList('website', 'js')
+        );
+    }
+
     public function testEntryPointsList_notFound()
     {
-        $extensionRuntime = new AssetRuntimeExtension($this->assets);
+        $extensionRuntime = new AssetRuntimeExtension($this->assets, new Router());
 
         $this->assertEquals([], $extensionRuntime->entryPointsList('fake'));
     }
@@ -211,8 +262,11 @@ class AssetRuntimeExtensionTest extends TestCase
 
         $assetRuntimeMock = $this->createPartialMock(
             AssetRuntimeExtension::class,
-            ['isHeadersSent', 'sendHeader', 'setCookie']
+            ['getRouter', 'isHeadersSent', 'sendHeader', 'setCookie']
         );
+        $assetRuntimeMock
+            ->method('getRouter')
+            ->willReturnCallback(fn() => new Router());
         $assetRuntimeMock
             ->method('isHeadersSent')
             ->willReturnCallback(fn() => false);